Move loadbalancer package to the coredns module
...as it's not used by the parent lighthouse module.
