+++ coredns/loadbalancer/smooth_weighted_round_robin_test.go
@@ -26,7 +26,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/submariner-io/lighthouse/pkg/loadbalancer"
+	"github.com/submariner-io/lighthouse/coredns/loadbalancer"
 )
 
 type server struct {
